Change user_id type from Int32 to Int64

+++ src/auth.cr
@@ -1,11 +1,11 @@
 module Auth
 
- def self.check_user_exists(user_id : Int32) : Int32?
+ def self.check_user_exists(user_id : Int64) : Int64?
    result = Db.connection.query_one?("SELECT id FROM users WHERE id = $1", user_id, as: {Int32})
    result
  end
 
- def self.register_user(user_id : Int32)
+ def self.register_user(user_id : Int64)
    Db.register_user(user_id)
  end
 end
